Adding ASOTracker
Currently, WMAgent has no idea of what ASO is doing. This basically
means throwing away/redoing a lot of code WMAgent already has written:

* DBS Publication (this was super hard to get right)
* Error handling/retry (this interacts deeply with the agent)
* Monitoring

Additionally, it doesn't make much sense conceptually to seperate out the
status of ASO from the rest of the agent. Currently the agent looks at
s FJR, and sees that the file is at site X, and everything's done. But
really, the file is (possibly) being moved to site Y, (possibly) has a new
LFN and (possibly) failed to even make it to the new place.

This patch make a new component (AsyncStageoutTracker) and adds a few new
states to the JSM to let it keep track of files that are moving. The
ASOTracker monitors ASO's user_monitoring database to figure out when files
have completed moving and them moves it to asocomplete/asofailed states
which can then be picked up by the JobAcountant where it's handled like
a normal job.

Since everything's contained in the agent, all the normal stuff works
right. You can publish to DBS, look at WMStats and have the error
algorithms fire the same way. You don't have to worry about combining
WMStats with the ASO database, and you don't have to worry about
maintaining a lot of old code.

One of the motivations for having the current architechture was
that ASO should be "decoupled" from WMAgent. I'd argue that this is
less coupled. Previously, it is a hack on top of a hack to, for
instance, figure out what job should be retried. A client would need to
have intimate knowledge of both ASO and WMAgent. This just requires
having the ASO user_monitoring database available. I'd argue that if
cmsweb's couch instance goes down, all bets are off anyway (since you
could no longer update WMStats, GlobalWorkqueue, etc..), and it's a
worthwhile tradeoff to remove a ton of glue code to keep two
seperate subsystems in sync

Conflicts:

	src/python/WMCore/WMSpec/StdSpecs/Analysis.py
	src/python/WMCore/WMSpec/StdSpecs/DataProcessing.py
	src/python/WMCore/WMSpec/StdSpecs/MonteCarlo.py
	src/python/WMCore/WMSpec/StdSpecs/StdBase.py